my_hello_flask/requirements.txt
jwt>=1.3.1
Flask-JSON>=0.3.4
psycopg2-binary>=2.9.1
bcrypt-3.2.0
